+++ featuretools/primitives/standard/rolling_transform_primitive.py
@@ -444,12 +444,33 @@ class RollingTrend(TransformPrimitive):
 
         Input datetimes should be monotonic.
 
+     Args:
+        window_length (int, string, optional): Specifies the amount of data included in each window.
+            If an integer is provided, will correspond to a number of rows. For data with a uniform sampling frequency,
+            for example of one day, the window_length will correspond to a period of time, in this case,
+            7 days for a window_length of 7.
+            If a string is provided, it must be one of pandas' offset alias strings ('1D', '1H', etc),
+            and it will indicate a length of time that each window should span.
+            The list of available offset aliases, can be found at
+            https://pandas.pydata.org/pandas-docs/stable/user_guide/timeseries.html#offset-aliases.
+            Defaults to 3.
+        gap (int, string, optional): Specifies a gap backwards from each instance before the
+            window of usable data begins. If an integer is provided, will correspond to a number of rows.
+            If a string is provided, it must be one of pandas' offset alias strings ('1D', '1H', etc),
+            and it will indicate a length of time between a target instance and the beginning of its window.
+            Defaults to 0, which will include the target instance in the window.
+        min_periods (int, optional): Minimum number of observations required for performing calculations
+            over the window. Can only be as large as window_length when window_length is an integer.
+            When window_length is an offset alias string, this limitation does not exist, but care should be taken
+            to not choose a min_periods that will always be larger than the number of observations in a window.
+            Defaults to 1.
+
     Examples:
         >>> import pandas as pd
         >>> rolling_trend = RollingTrend()
         >>> times = pd.date_range(start="2019-01-01", freq="1D", periods=10)
         >>> rolling_trend(times, [1, 2, 4, 8, 16, 24, 48, 96, 192, 384]).tolist()
-        [nan, nan, 1.5, 3, 6, 8, 16, 36, 72, 144]
+        [nan, nan, 1.4999999999999998, 2.9999999999999996, 5.999999999999999, 7.999999999999999, 16, 36, 72, 144]
 
         We can also control the gap before the rolling calculation.
 
@@ -489,7 +510,7 @@ class RollingTrend(TransformPrimitive):
                 x, self.window_length, gap=self.gap, min_periods=self.min_periods
             )
             if isinstance(self.gap, str):
-                additional_args = (self.gap, np.mean, self.min_periods)
+                additional_args = (self.gap, calculate_trend, self.min_periods)
                 return rolled_series.apply(
                     apply_roll_with_offset_gap,
                     args=additional_args,
